Suurem osa vana pathfinderiga seotud funktsioone kustutatud, hetkel nexust valides kalkuleerib programm tee, näitab seda graafiliset ning edasi ei toimu midagi.

+++ src/NewPathfinder.java
@@ -115,7 +115,7 @@ public class NewPathfinder {
         Canvas canvas = Game.getCanvas();
         GraphicsContext g = canvas.getGraphicsContext2D();
         for (Node node : visited) {
-            String cost = Double.toString(node.getCost());
+            String cost = Integer.toString((int) node.getCost());
             int centreX = Game.indexToPixel(node.getX());
             int centreY = Game.indexToPixel(node.getY());
             g.setFont(Font.font("Calibri", FontWeight.BOLD, 10));
